refactor: KDA-49 패키지 구조 변경
