+++ presentation/src/main/java/com/upf464/koonsdiary/presentation/ui/account/signup/EmailSignUpViewModel.kt
@@ -1,4 +1,4 @@
-package com.upf464.koonsdiary.presentation.ui.account
+package com.upf464.koonsdiary.presentation.ui.account.signup
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
